requirements.txt
customtkinter==5.2.2
discogs_client==2.3.0
matplotlib==3.9.2
numpy==2.1.1
pandas==2.2.3
Pillow==10.4.0
pymongo==4.9.1
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.66.2
